WEAV-1353 (P1) -- Add connection support to Binding WEAV-268 (P2) -- Supply optional "Binding Complete" handler for the application.
-- Added the ability to initiate TCP connections using the Weave Binding API.  When Transport_TCP() is configured the
binding will automatically initiate a TCP connection to the specified peer address.  If Security_CASESession() is also
specified, a connection-bound CASE session will be negotiated once the connection is established.

-- Added an Binding API event (ConnectionEstablished) that is delivered to the application upon a successful TCP connect.
This gives the application an opportunity to adjust the configuration of the connection, e.g. to enable TCP keep-alive.

-- Added the ability to prepare a binding using an existing Weave connection via the Transport_ExistingConnection()
configuration function.

-- Exposed reference counting APIs for WeaveConnections and implemented support for ‘lazy closing’ of connections upon
release of the last reference.

-- Modified Weave Binding and ExchangeContext objects to hold references to WeaveConnections and release them when the
holding object closes.  This supports a pattern of automatic connection establishment and termination without the need for
explicit connection management by the application.