Send Tunnel Liveness when there has been no outbound message for Timeout period regardless of any receipt of inbound message.
The liveness timeout, as it is implemented currently at the Service
is composed of a separate read and write timeout that fire independently
to close the connection if there is no traffic in the corresponding direction.
In the absence of the TCP UserTimeout implementation at the Services,
this helps to failover within 200 seconds on the Services side when
there is no traffic from the device even though the Service is sending traffic
down the potentially dead connection.
However, this creates a discrepancy in the Tunnel liveness mechanism between
the device and the Service.
One way to remove this discrepancy is to relax the liveness strategy on the device
side and have it send a liveness message if it has not sent a message up in 200
seconds regardless of whether it has received any messages or not.

By virtue of the symmetric nature of the tunnel liveness message,
this would effectively help in feed both the timers at the
Services' side and prevent spurious closing from their end.