WEAV-2323 (P1) -- Binding: Add support for establishing non-shared CASE sessions WEAV-2324 (P1) -- Security: Manage lifetime of non-connection bound sessions
-- Added support for establishing non-shared CASE sessions via Bindings using the Security_CASESession()
configuration method.  As Bindings do not yet support TCP, this option is only available for bare UDP and UDP with
WRMP.

-- Implemented a simple queuing mechanism that allows binding prepare operations that involve establishing new
security sessions to queue when the Weave security manager is busy, rather than failing with a SECURITY_MANAGER_BUSY
error.  This eliminates the need for retry logic at the application level.

-- Added a scheme for encryption key reservation and release.  Key reservations are used by Bindings and
ExchangeContexts to signal to the security manager that a key is actively in use and should be retained.  Although
the reservation APIs are called for all key types, currently reservations are only tracked for session keys.  This
works in conjunction with the session key removal features described below.

-- Implemented behavior in the security manager to automatically remove non-connection bound (UDP) session keys
whenever their reservation count goes to zero.  This has the effect of terminating a UDP-based session once the
Binding that created it and any associated ExchangeContexts are gone.  This behavior does *not* apply to session
keys that are bound to TCP or WoBLE connections which live until the connection ends.  Additionally, the
‘remove-on-idle’ feature (see below) can be used to extend the lifetime of a session even when it’s not being used.

-- Added a facility to automatically remove session keys after a period of idleness.  The ‘remove-on-idle’ feature
only applies to non-connection bound (UDP) sessions, and only when the session is specifically marked for
remove-on-idle.  Currently all remotely-initiated sessions are marked remove-on-idle at session establishment time.

-- Added an API to the security manager to abort an in-progress session establishment.  This is used by Bindings
when the user calls Close() during the prepare security phase.