Implement resetting of the Tunnel Reconnect backoff.
Surface a ResetReconnectBackoff as an API of the WeaveTunnelAgent
that can be requested with a boolean argument to perform the subsequent
reconnect immediately or after waiting a random period within a
configured time window.
This API would be used to effect an accelerated Tunnel recovery based
on positive information about underlying network connectivity.

Also, added 2 test-cases to the Tunnel test suite, 1 for the immediate
reconnect and the other for the delayed one.