Implemented WEAV-2335 (P1) -- Binding: Add DNS support
-- Modified Weave Binding class to support DNS.  The Binding::Configuration class now offers two new versions
of TargetAddress_IP() that take a string host name for the peer (one NULL terminated, one with an explicit
length).  If DNS name resolution is enabled (WEAVE_CONFIG_ENABLE_DNS_RESOLVER == 1), the Binding will use DNS
to resolve the peer’s address.  If DNS is not enabled, but WEAVE_CONFIG_RESOLVE_IPADDR_LITERAL == 1, the
Binding will attempt to parse the supplied string as a string-form IP address.

-- Enhanced TestBinding tool to support testing Binding DNS functionality.