Fix for WEAV-2332.
Add logic for parsing a StatusReport in response to Tunnel control
messages with TLVData in it.

If present with the restricted routing tag set, then border routing
would be disabled by not installing the fabric tunnel route.
Additionally, pass an appropriate error code to the application
that indicates this operational mode.

Add a test-case in the Tunnel test suite.