Fixed time sync bugs and improved tests
The tests now check that the sync succeded and that the source is
realiable.
Strace is now disabled by default.
There are now three .sh files to make it more convenient to execute the
three flavors of sync test implemented in test_weave_time_01.py.
WEAV-2342: Fixed TimeSyncNode::RegisterCorrectionFromServerOrNtp and
RegisterLocalSyncOperation so they can be called from callbacks;
this lets the Coordinator register the timestamp of the latest good
sync when it gets an upcall from the Client code.
WEAV-2343: Fixed TimeSyncNode::EndLocalSyncAndTryCalculateTimeFix so
when it ignores an update because the clocks have not drifted enough,
the reliability of the time source is not discarded.
WEAV-2345: Fixed TimeSyncNode::EndServiceSyncAndTryCalculateTimeFix so
it does not ignore the Freshness value sent by the server.