Limit the max amount of plaid-speed time in tunnel tests
The tunnel tests need the amount time run at plaid speed to be limited
to avoid the gateway to race ahead and process more events while the
test script is sending it a SIGUSR1 to trigger a cleanup.
Without a limit on the amount of time spent at plaid-speed, the gateway
has time to process extra keepalive messages that make the test not
consistent with an execution at realtime speed.