Eliminate clearing of characteristics until later in the discovery process
For whatever reason, CoreBluetooth ends up generating two callbacks
for didConnectPeripheral.  This is the real issue and one that remains
unsolved here.  Connecting the peripheral twice without an
interviening disconnect caused us to discover the services and
characteristics twice as well.  This exposed a race condition: if we
used the WoBLE API during the second discovery process we would hit an
empty characteristics list. The race condition did not surface in the
commandline WDM, but did surface in the more intense exercise of the
API.  The workaround implemented here does not invalidate the
characteristics list until a new list is fully generated and then it
attempts to do it via a single unconditional assignment.