RAM usage optimizations
Reduced the size of several fields of the Record structure from uin32_t
to uint16_t or uint8_t. The APIs still take uint32_t, so the internal
representation can be made configurable in the future if larger counters
are required on some platforms.
Updated the API to accept const arrays of pointers to const strings,
so the application can keep both the strings and the pointers to the
strings in read-only memory.