Add a preprocessor to the doxygen pipeline
WeaveMakeManagedNamespaceIdentifier was causing doxygen not to process
correctly any of the documentation for the WDM Next.  This change
creates a filter that expands that construct into something more
usable, and generates the appropriate documentation.  I have not
managed to achieve the same via Doxygen's built-in
ENABLE_MACROS/PREDEFINED settings, so the simple shell script was the
next option.