Fix for WEAV-2599
Make a call to the WARM TunnelModeChange API instead of the
ServiceTunnelEstablished API when a second tunnel is established.

The issue was that the TunnelAgent was calling the latter API with
the right availability mode which was not changing the route
priorities when Primary tunnel came back up after
Backup was already up.

Also, fix coverity issue at
https://coverity.nestlabs.com/query/defects.htm?projectId=10002&cid=55718