Fix for WEAVE-2556 (P1) -- Lost key confirmation can strand a binding in a bad state
-- Modified the behavior of the Weave Security Manager such that it rejects any
concurrent requests for a shared CASE session when the session exists in the session
table but the security manager still considers session establishment to be incomplete.

This situation can arise when establishing CASE over Weave Reliable Messaging.  After
the initiator has sent a KeyConfirm message it waits for a WRM ACK from the responder.
During this time, the session exists in the session table but is not yet considered
ready for use.  Until the ACK is received, additional requests to establish the same
shared session should be denied with a SECURITY_MANAGER_BUSY error, which will force the
concurrent request to wait until the session is fully established.