Various changes to Weave Binding, big and small
-- Added support for initiating PASE sessions.  By default the device pairing code is used
as the password.  However this can be overridden by the application via a Binding API event.
Due to limitations in the security manager this feature can only be used over connection-
oriented transports.

-- Added support for initiating TAKE sessions.  Because TAKE initiation requires a client-
side delegate object (WeaveTAKEChallengerAuthDelegate), along with other protocol-specific
configuration parameters, a new API event was added to the Binding API to request these
parameters from the application at the moment of initiation.  This pattern eliminates the
space required to store these values within the Binding object itself.  Like PASE, TAKE can
only be used over connection-oriented transports.

-- Modified the Binding PrepareFailed API event such that it passes StatusReport information
to the application when preparation fails due to the peer returning a non-successful status
report.  This enables better error reporting in the application.  Currently the only
situation in which a StatusReport information is returned is when a security session fails
with an error from the peer.  (This closes a long-standing TODO).

-- Improved logging when binding preparation fails due to receipt of a status report.  Now
the log message includes the status profile and code from the StatusReport and a
corresponding string translation if available.

-- Modified WeaveExchangeManager to notify each Binding object when a connection is closed.
This was the intended design when Bindings were enhanced to support connections.  However
somehow the behavior was never fully implemented.

-- Modified Binding::PrepareTransport() to clear the default connection closed handler (as
installed by the message layer) when instantiating a new connection. This is needed to avoid
a double release of the connection on a network close.

-- Fixed a WeaveConnection leak.  When an ExchangeContext is vended from a Binding that has
been configured to use a WeaveConnection, a reference to the connection it taken to ensure it
remains live for the duration of the exchange.  The ExchangeContext is expected to release this
reference when it is no longer needed.  However, in the event the connection experiences a
network close, code in WeaveExchangeManager::HandleConnectionClosed() was not releasing this
reference before discarding the pointer to the connection, resulting in a leak.

-- Added GetPeerIPAddress() method to allow applications to retrieve the IP addressing
information (address, port and interface) for the peer.

-- Added GetExchangeManager() method.

-- Added #if statements to reduce code size in cases where CASE/PASE/TAKE initiation is
disabled at compile time.