New Weave Echo profile APIs
-- Introduced new WeaveEchoClient and WeaveEchoServer classes as replacements for
the venerable versions introduced back in February 2013 (5 years ago!). Both
classes employ the modern event-based API pattern.  The new WeaveEchoClient class
uses Bindings to establish communication with a peer and offers features such as a
built-in repeating send mode and support for multicast.  WeaveEchoServer now gives
the application better control over the generation of Echo Responses, including
the ability delay or suppress responses and alter their payloads.

For the time being the original classes are maintained as the default API for
backwards compatibility.  The new APIs are available for use by including the
Profiles/echo/Next/WeaveEcho.h header or defining WEAVE_CONFIG_ECHO_NAMESPACE to
kWeaveManagedNamespace_Next.

As of this commit there is no code in the source tree that uses the new API.
However the new code has had extensive manual testing using a new version of
weave-ping that will be released once feature parity is reached with the current
version.