ExchangeContext: explicitly cast enum to uint16_t
Newer compilers and C++ versions are increasingly picky about "narrowing"
conversions. And meanwhile, the underlying type for an enum is (by default) an
implementation-specific detail of the compiler. (However, in C++11 you may
explicitly declare the underlying type of an enum.)

Thus, with GCC 6 type deduction fails for GetFlag() and SetFlag() because it
won't implicitly convert an enum to uint16_t. Therefore, adding an explicit cast
to the code.

Signed-off-by: Gabriel M. Beddingfield <beddingfield@nestlabs.com>