fix segmentation fault in wobluez manager
The segmentation fault happen in scan loop, because shutdown or exit can
hang with daemon thread running. Use the workaround that use for loop to check if specific peripheral can be found, and remove thread close in the end.

This looks like a python bug.

Reference bug: https://bugs.python.org/issue1856