Generalize ResourceIdentifier
Prior to this commit, ResourceIdentifiers were only really supported
as uint64_t that were meant to correspond to device IDs.  There was
some initial sketch work pointing towards ResourceIdentifiers being
16-byte UUIDs.  With this commit, we generalize the concept of a
ResourceIdentifier, and we centralize it in a single place.  At the
moment, the following resource ID formats are supported:

* generalized ID.  length is 10 octets, with the first two octets
  encoding the ResourceType in little endian, and the remaining eight
  octets encoding the ResourceId in little endian.

* plain uint64_t in little endian.  This representation will generate
  the generalized representation of the ResourceIdentifier with the
  ResourceType being set to the agreed upon value (1, corresponding to
  RESOURCE_TYPE_DEVICE in weave/common/resource_type.proto)

The ResourceIdentifier embodies the commonly used encoding
conventions: the default constructor generates a resource id that
corresponds to the device/self node ID, that subsequently gets
compressed out of the representation. TLV serialization and
deserialization routines are provided.

Change-Id: I2e2821d815d3dd77e954ee46b50fc586903002a9