Fix the filter algorithm when DE version is equal to sink trait version
somehow device become offline, local changes has been made via SetUpdated, assume no changes happen in service side, when
it resubscribe, the version in notify may be same as local version for
related trait,
a. if the trait is only for unconditional, in this situation, local changes in pending set  wins.
b. if the trait is only for conditional, in this situation, local
changes in pending set or dispatched set wins.