Added SetVersion() Protected Method to the TraitDataSink Class.
It is needed because WDM clients should be able to persist version
value and initialize it when the value is lost in the volatile memory.
The new method is protected so only the implementation has the privilege
of overriding the version.

For consistency, added similar method to the TraitDataSource Class and
made mVersion aprivate member.

Problem(s):
    https://jira.nestlabs.com/browse/WEAV-2703