Fix for SUN-7924.
As part of changes for COM-3004, the OnlineChecker was
hooked on to the WeaveTunnel for periodically checking if broader
network connectivity was restored so that the Tunnel could perform
an accelerated reconnect to expedite Tunnel recovery and counter
the slowdown introduced by its backoff.

This reconnect attempt ran into a race condition with the regular Tunnel
connect attempt and produced a scenario where the Tunnel connection was
closed(to reconnect again) when a TunnelOpen was already sent.
A method that attempted to release resources(e.g., Tunnel control
ExchangeContext) found it already closed by the TunnelOpen response
expiry handler.

The fix, essentially, prevents a ResetReconnectBackoff to close an
existing connection and reconnect if an attempt to establish the
Tunnel is already underway.

Added test case for trying to reset the reconnect backoff after a TunnelOpen
has been sent and verifying that it does not cause any disruption to
that attempt.