Project Blue Sky : Support for device account pairing
-- The ServiceProvisioningServer now supports account pairing by sending the
PairDeviceToAccount request to the service over the service tunnel.  Only
service tunnel pairing is supported in the current code.  The new code
automatically handles waiting for service connectivity to be established
before attempting to send the PairDeviceToAccount request.

-- Separated the states of a device with respect to being service provisioned
and being paired to an account.  A device is service provisioned when it has a
persisted service configuration.  Whereas a device is paired to an account once
PairDeviceToAccount completes and the account id has been persisted.

-- Introduced a new AccountPairingChange event which fires when PairDeviceToAccount
completes, or when UnregisterService is invoked.

-- Added new platform event for alerting the application, and other platform
components, to changes in time synchronization.  Currently, only transitions
from unsynchronized to synchronized are noted.

-- Modified the IsPairedToAccount() virtual methods on the provisioning server
objects to check the new “is paired to account” state maintained by the
Configuration Manager.  The effect of this is to enforce the "paired device"
access control rules only once the device has fully completed pairing to an
account.

-- Added a build-time option to disable account pairing.  When this option is
enabled, the device performs all local operations associated with registering
a service, but will not send the PairDeviceToAccount request to the service.
This is intended to facilitate testing.