Project Blue Sky : Various changes to Connectivity Manager
-- Fixed a bug introduced by a change in the behavior of the ESP WiFi layer.
The WiFi layer no longer supports a mode where neither the station nor the AP
interfaces are enabled.  Because of this the Connectivity Manager manager
wasn’t able to properly control the state of the AP interface, resulting in
some confusion between the layers.  This required a rather significant change
to both the DriveAPState() and DriveStationState() methods.

-- Moved various utility functions related to interacting with ESP APIs into
a new ESPUtils class.

-- Added an interface to ConnectivityManager to query the state of the AP
interface.