Moving BLE adapter powerup functionality to adapter proxy enumeration callback, JIRA:MS-1132
Seems like there is a race condition between WeaveClientReady() &
adapter proxy enumeration.  Current code tries to power up the adapter
in WeaveClientReady(). But, there is possibility that adapter has not
been enumerated when this function is called. I think this is causing
the crashes reported in MS-1132.  Now, as adapter power up is
happening in its proxy enumeration callback, above scenario will not
occur.