Randomized Trait Data Versions
This adds support for randomized trait data versions to
TraitDataSources.

This utilizes RandUtils::GetRandU64 to populate the version on creation
of the data source. GetRandU64() ultimately calls rand(), and
the latter cannot be invoked from within statically constructed objects
(of which there are quite a few on FreeRTOS based platforms).

To work around that, the version is populated on first observation or
mutation of the private version member variable, circumventing the need
to seed it at construction.

Issue(s):
        WEAV-2611

Test(s) Run:
        Added support to the TestTDM unit to test for ranodmized
        versions.