Project Blue Sky : Various changes to configuration
-- Changed the Device Vendor Id and Device Product Id config setting to accept
values in hex.  This is more inline with the usual way these values are depicted.

-- Changed the default test pairing code “NESTUS”.  This value has the benefit
of being fairly memorable while also including a valid Verheoff check digit.

-- Added an sdkconfig.defaults file which includes default configuration values
for ESP-IDF components to be used when the application is initially configured.
Current this file contains settings that enable BLE server support by default.