Added pairing code utility functions to WeaveSecuritySupport Java Libaray
-- Added a new PairingCodeSupport class to the WeaveSecuritySupport Java library.  This class
provides static utility methods for working with pairing codes, including methods for generating
pairing codes, validating them and normalizing them.

Some of the functionality of PairingCodeSupport overlaps that provided by the existing
nl.Weave.DeviceManager.PairingCodeUtils class.  The intent is that the new class will ultimately
supplant PairingCodeUtils and provide a more consistent and maintainable interface to this
functionality for Java and Android applications.

-- Added a C++ function, GeneratePairingCode(), that can be used to generate random pairing codes of
arbitrary length.

-- Implemented unit tests for the above.

-- Updated make check target to run security support tests.