Notify application when the send channel of the TCP connection changes state between being idle or not-idle.
This feature is configurable at compile time on the TCPEndPoint.
When enabled, it does the following:
After sending data, the device polls the TCP send queue periodically
and if it has flushed all pending data, the application is notified
and polling stopped.

If the data flow has not progressed and the configured TCP UserTimeout
expires, the connection is closed.

If the data flow progresses at any poll, the TCP UserTimeout timer
is reset.

The TunnelAgent utilizes this TCPEndPoint feature to, subsequently,
notify the application.

Add Tunnel test case for sent-data flushed notification.