Retargetable Device Layer : Refactor PlatformManagerImpl
-- Refactored PlatformManagerImpl into two generic implementation templates:
GenericPlatformManagerImpl<> and GenericPlatformManagerImpl_FreeRTOS<>.  These
templates provide reusable implementations of the PlatformManager abstract
interface that are suitable for use on all platforms and FreeRTOS platforms,
respectively.

-- Re-implemented the ESP32 PlatformManagerImpl class using
GenericPlatformManagerImpl_FreeRTOS<> as a base.  This resulted in a significant
reduction in the amount of code that is specific to the ESP32 implementation.

-- Moved the implementations of the System Layer event and timer platform
functions into separate files and restructured the code such that these methods
simply delegate their functionality to PlatformManager internal methods.

-- Moved the ESP32-specific LwIP core locking code into its own, platform-
specific file.

-- Renamed the file containing ESP32-specific time/clock functions.