WeaveCASEAuthDelegate Enhancements
-- Modified the WeaveCASEAuthDelegate API to support platforms where access to the local node’s
private key is restricted to a hardware element or a secure enclave.  With the new API, the
operation of signing a CASE message (or more specifically, signing its hash) is forwarded to the
delegate object, which is expected to compute and encode the signature synchronously before
returning.  This pattern is in contrast to the previous design where the CASEEngine queried the
private key and perform the signing operation itself.

-- Revised the WeaveCASEAuthDelegate API to provide additional context information to each of the
delegated method calls.  The context information provides delegate implementations will full
access to the information from the CASE message being processes.  This can be used, for example,
to customize certificate validation policy based on information about the peer.

-- Implemented a compile-time compatibility mode that provides support for the legacy
WeaveCASEAuthDelegate API.  This mode is enabled via the WEAVE_CONFIG_LEGACY_CASE_AUTH_DELEGATE
configuration option.

The legacy mode is enabled by default in WeaveConfig.h, but *disabled* in stand-alone builds (in
build/config/standalone/WeaveProjectConfig.h).  This provides backwards compatibility with
existing platform implementations while allowing testing of the new API in desktop builds.

-- Revised the CASE auth delegate implementations that exist in the Weave Device Manager and the
test applications to support both the new API and the legacy API.