WeaveKeyExportDelegate Enhancements
-- Revised the WeaveKeyExportDelegate API to support platforms where access to the local node’s
private key is restricted to a hardware element or a secure enclave.  This is a parallel change
to the recent WeaveCASEAuthDelegate change.

-- Restructured the WeaveKeyExportDelegate API such that the delegate methods take a pointer to
the key export engine object.  This simplifies the means by which context information is passed
to the delegate.

-- Cleaned-up the API of the WeaveKeyExport object by making some members private and adding
accessor methods.  Also, moved WeaveMessageInfo context pointer into the WeaveKeyExport object
so that it doesn’t have to be passed on each delegate API call.

-- Added a compile-time compatibility mode to the delegate object that supports the legacy
WeaveKeyExportDelegate API.  This mode is enabled via the WEAVE_CONFIG_LEGACY_KEY_EXPORT_DELEGATE
configuration option.

The legacy mode is enabled by default to provide compatibility for existing platforms.  For
stand-alone builds the new API is enabled to allow for testing.

-- Replaced the various GenerateWeaveSignature() public functions with a set of WeaveSignatureGenerator
classes whose methods provide similar functionality. These classes rely on virtual methods calls
to compute the low-level signature data, which allows an API client (in this case the key export
engine) to own the signing operation.

-- Updated various instances of key export delegate objects to support the new API.  In all cases
the code supports both the new API and the legacy API.