Additional support for EAX mode message encryption
-- Implemented key derivation for EAX sessions in WeaveCASEEngine. (NOTE: Equivalent support
for PASE, TAKE and group keys is TBD).

-- Enhanced WeaveSecurityManager and Weave Binding object to allow encryption type to be
specified when initiating a CASE session.  (Again, equivalent code for other code paths is
TBD).

-- Added a command line option to the test apps for specifying the message encryption mode.
The --msg-enc-type option accepts one of three arguments: aes-128-ctr-sha1 (the default),
aes-128-eax-128 and aes-128-eax-64.  Note that, in the current version of the code, only the
weave-ping and TestBinding tools honor this option.

-- Re-wrote the TestMsgEnc tool to include test coverage for the new EAX message encryption
formats.  The new version of the tool also provides broader test coverage for existing message
formats (both encrypted and unencrypted).  An accompanying script is used to generate a set of
message encoding test vectors (575 in total) which are used by the TestMsgEnc tool to verify
the correct behavior of the Weave Message Layer. The script uses a parallel implementation of
Weave message encoding constructed in pure Python to provide an independent verification of
the behavior of the C++ code.

-- Added a simplistic message decode fuzz test to the TestMsgEnc tool.  This test works by
mutating each bit in a message, one at a time, and verifying sane behavior of the message
decode function.  Fuzzing is performed for each of the test vector messages described above.