Changes to Weave message integrity check code
-- Modified the code that generates the message integrity check pseudo-header
for EAX-encrypted messages such that the message header and id are always included
in the pseudo-header, even for V1-format messages.  This is in contrast to
messages encrypted with AES128CTRSHA1, which for reasons of backwards compatibility
do not incorporate these fields.

-- Collapsed the code for generating the integrity check pseudo-header into a
single function used for all encryption types.

-- Made minor documentation and editorial changes to the code surrounding the
message integrity check computations.