Service Fabric Join message definition for In-field Joining(IFJ) in Service Provisioning profile.
Legacy devices in the field would employ IFJ to join the Weave
fabric of the Structure so that they can communicate using Weave
with the rest of the Weave devices in the home.
This message would be used by that device, at the end of IFJ, to
signal successful completion of the process to the Service.
After this communication, the device is formally part of the
Weave fabric of the structure and can assist other new device
pairings as well.