Small additional check for logging external events
Add a check to ensure that the `EnsureSpace()` completes successfully.
This is largely a cosmetic change, as the space reserved in that
particular call should be smaller than WEAVE_CONFIG_EVENT_SIZE_RESERVE
for just about any case, but checking the return value and taking
appropriate action will not cause any adverse runtime effects and will
minimize the code maintenance in the future.