Only restart WRMP timer if needed
This change reduces flooding of the Weave tasks' message queue on
NLER-based platforms in some circumstances. In NLER if a timer has
expired and posted its timer event but that event has not yet been
processed by the destination task (Weave task), and that timer is
then cancelled, the event will remain in the destination task's event
queue until the destination task dequeues it (and subsequently ignores
it, as it has been cancelled). This can lead to flooding of the message
queue by timer messages if we experince a burst of incoming traffic that
rapidly toggles the WRMP timer on and off. This is especially bad in
cases where we are overdue for a WRMP timer expiry, when every timer
toggle is an immediate expiry.