virtual-ap.sh -- Tool to setup a virtual WiFi AP on linux
virtual-ap.sh is a shell script that automates the process of setting up a Virtual WiFi
access point on a linux system using an available WiFi interface.  virtual-ap.sh is
intended to support the development and testing of WiFi-enabled IoT devices without the
need for a dedicated access point.

virtual-ap.sh uses hostapd, dnsmasq and iptables to create a WiFi access point that
provides DHCP and DNS services for attached clients.  Clients are granted access to
external destinations via the host's primary interface, but only those with public IPv4
addresses; access to destinations with private (RFC-1918) addresses is blocked to
improve security.  Access to the linux host itself is also blocked, with the single
exception of the Weave port (11095).

Various command line options provide the ability to tune the AP configuration as needed.
Additionally, the script is capable of orchestrating a graceful shutdown of the AP,
leaving the host system in its original state.

More work on virtual-ap.sh

-- Added missing sudo when enabling IPv4 forwarding.

-- Added code to automatically disable management of AP interface on systems
that use Network Manager.

-- Made minor clean-ups to the structure of the code.

More work on virtual-ap.sh

-- Show an abbreviated help message when run without specifying an operation.