Work in Progress : New adaptation of OpenWeave to Android
-- This is an initial cut at a new adaptation of OpenWeave to Android platforms, with an emphasis on
supporting Android mobile development.  The primary aim of this commiot is to establish a new build
process for OpenWeave code that better aligns with traditional Android build tools and
methodologies.

As an initial proof-of-concept, the code builds a version of the Weave Security Support library for
Android. The Security Support library is a simple Java wrapper for various stand-alone cryptographic
and security-related functions (e.g. converting a certificate between Weave and X.509 format).  This
library was chosen for its simplicity, allowing the adaptation effort to focus on build issues
rather than the particular code being built.

Unlike the existing top-level Makefile-Android, this new adaptation uses gradle and a build.gradle
script to orchestrate the build.  This in turn calls out to a helper makefile (BuildWeave.mk) to
drive building the OpenWeave C++ code via the traditional automake scheme.

Also unlike Makefile-Android, build.gradle produces a single Android Archive (AAR) file containing
shared libraries for all ABI targets.  This is in contrast to Makefile-Android which builds a JAR
file containing a single shared library for one target.  This makes the output easy to drop into
standard Android applications.

The build.gradle script relies on the Android Gradle plugin (albeit, a slightly old version in the
current code).  The build can be invoked from the command line, using the supplied gradle wrapper
script, or it can be imported into Android Studio and built from there.

Quick build instructions:

- Install a version of the Java JDK (openjdk-11-jdk is known to work)

- Install the Android SDK for Linux

- Create a local.properties file in the src/adaptations/android/weave-security-support directory
containing the paths to the SDK and NDK directories.  E.g.:

    ndk.dir=/home/jay/tools/android-sdk-linux/ndk-bundle
    sdk.dir=/home/jay/tools/android-sdk-linux

- Invoke the gradle wrapper to build the library.

    ./gradlew build