WIP: Device layer port to nRF52840
-- Initial working integration of OpenWeave stack into demo application.  Init and
configuration of OpenThread is hard-coded in the demo app.  ICMP6 works, but inbound
packets from thread interface are simply discarded (i.e. there is no integration with
LwIP yet).

NOTE: Entropy implementation in DeviceLayer is purposefully broken due to incompatibility
with OpenThread's use of the nRF52840 entropy source.  This remains to be fixed.

-- Fixed small logic bug in BLEManagerImpl::_SetAdvertisingEnabled().