Add __APPLE_USE_RFC_3542 to ensure IPV6_PKTINFO is available
In an earlier revision of the Inet layer, all code that relied on IPV6_PKTINFO had
the above constant defined.  When chunks of code moved from UDPEndPoint to IPEndPointBasis,
the above define was missing which resulted in code that depended on IPV6_PKTINFO being
unavailable.  In turn, that made it impossible to successfully perform a rendezdous on
a MacOS platform using link local addresses.

We are opting to put the define in the file directly rather than in the WeaveProjectConfig.h
for Darwin; the approach we took is more robust against changing configuration files