nRF52840 adaptation: Continued work on OpenThread integration
-- Added initial support for OpenThread within the OpenWeave Device Layer.  The Device
Layer now exposes a ThreadStackManager API that provides platform-neutral methods for
initializing and operating a Thread network stack from within a Weave-enabled application.
A single, concrete implementation of the ThreadStackManager API (ThreadStackManagerImpl)
is provided which supports the use of the  OpenThread stack within the context of
FreeRTOS, LwIP and the nRF52 SDK.

-- Introduced 3 new generic implementation classes that implement key features associated
with the ThreadStackManager API:

GenericThreadStackManagerImpl_FreeRTOS<> - Provides a generic implementation of
ThreadStackManager features for use on FreeRTOS platforms.

GenericThreadStackManagerImpl_OpenThread<> - Provides a generic implementation of
ThreadStackManager features for use on platforms that use OpenThread.

GenericThreadStackManagerImpl_LwIP<> - Provides a generic implementation of
ThreadStackManager features for use on platforms that use OpenThread in combination with
LwIP.

-- Added support for queuing OpenThread state change events in the Weave Device Layer
event queue.

-- Made LwIP interface names for the WiFi, Thread and tunnel interfaces configurable at
compile time.

-- Add support for the Thread interface to the WARM function used to retrieve an LwIP
netif given a WARM interface type.

-- Fixed bug in GenericPlatformManagerImpl_FreeRTOS<>::_RunEventLoop() which resulted
in events being dispatched without the stack lock being held if event were already in
the queue when the method was called.