nRF52840 adaptation: Restructed ConnectivityManager implementation classes
-- Added new GenericConnectivityManagerImpl classes that provide minimal default
implementations for platforms without certain features.  For example, the
GenericConnectivityManagerImpl_NoWiFi<> class provides minimal implementations
for the WiFi-specific features of the ConnectivityManager on platforms that don’t
have a WiFi interface.  These generic implementations can be mixed into a
platform’s ConnectivityManagerImpl class to avoid the need to for each platform
to define methods for features that aren’t available.

Similar “NoBLE” and “NoTunnel” classes are provided for the BLE and service
tunnel-specific features of ConnectivityManager.

-- Added a GenericConnectivityManagerImpl_BLE<> class, which provides
implementations of the BLE-specific ConnectivityManager features that forward the
method calls onto the BLEManager object.  For platforms that support BLE, this is
the standard way of implementing these methods.  Thus this class eliminates
repetitive code across different platforms.

-- Modified the ConnectivityManagerImpl for the nRF52 platforms to use the
GenericConnectivityManagerImpl_BLE<> or GenericConnectivityManagerImpl_NoBLE<>
generic implementation classes depending on whether support for WoBLE is enabled.

-- Modified GenericThreadStackManagerImpl_LwIP<>::UpdateNetIfAddresses() to ignore
Thread RLOC addresses when populating LwIP’s address table.