nRF52840 adaptation: Continued work on WARM, logging
-- Completed integration on WARM for devices with Thread interfaces.  Weave
addresses are now automatically assigned / removed from the Thread interface
whenever a device’s membership in a Weave fabric changes.

-- Updated logic within GenericThreadStackManagerImpl_LwIP<> to toggle the
link state of the LwIP Thread interface based on the current Thread device
role (more specifically, whether the device is attached to a Thread network).
This same code also posts a new ThreadConnectivityChange event on transitions
in the link state.

-- Made significant enhancements to debug logging related to events in
OpenThread and the LwIP Thread interface.

-- Minor edits to the README.