Corrections to LwIP's IPv6 source address selection algorithm
-- LwIP's IPv6 source address selection algorithm was modified to more
closely match the rules in RFC 6724; in particular, rule 8 which governs
the partial matching of source addresses against a destination address.

Previously, LwIP would only prefer a source address that partially matched
the destination address if the two had a /64 prefix in commmon. With the
new code, LwIP will now prefer a source address that has a /48 prefix in
common if no address with a longer common prefix exists.

This behavior is still not completely faithful to RFC 6724, which specifies
selecting a source address with the absolute longest common prefix. However,
the new code is sufficient to properly select source addresses in networks
with multiple /16 subnets, which is a common use case, especially when using
with ULAs.